Reorganize projects component
